move map interactions to its own hook
